CPF-179 Make StandardJsonFormat Zeitwerk-compatible
